Added object sieving to type lists & added a pyimgui interface

+++ ValkyrieDLL/ValkyrieDLL/GameData.cpp
@@ -259,7 +259,7 @@ void GameData::LoadUnits(const char* fileName, float& percentValue, float percen
 	}
 }
 
-bool LoadTextureFromHeap(void* heap, int heapSize, PDIRECT3DTEXTURE9* outTexture) {
+bool LoadTextureFromHeap(void* heap, size_t heapSize, PDIRECT3DTEXTURE9* outTexture) {
 	PDIRECT3DTEXTURE9 texture;
 
 	Valkyrie::DxDeviceMutex.lock();
